backend/pyproject.toml
[project]
name = "koodfood"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.9.1",
    "celery>=5.5.3",
    "coverage==7.10.7",
    "dj-database-url>=3.0.1",
    "django==5.2.5",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.29.0",
    "environs>=14.3.0",
    "factory-boy==3.3.3",
    "faker==37.8.0",
    "flower>=2.0.1",
    "gunicorn>=23.0.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "psycopg2-binary==2.9.10",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
    "whitenoise>=6.11.0",
]
